requirements.txt
aiohttp
astroid
async-timeout
attrs
chardet
colorama
discord.py
idna
isort
lazy-object-proxy
mccabe
multidict
pip
pylint
setuptools
six
toml
websockets
wrapt
yarl